requirements.txt
asgiref==3.5.2
coverage==6.5.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==5.0.4
mccabe==0.7.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
